Replace 'acquire_script' with 'post_autoreduction_job'
